emit: fix race condition warning
The whole code is a hack, as I clearly just copy-pasted huge block of
code into another internal module and found a hacky method to re-use it.
Then I needed to validate that emitter exists and only way was to send a
test message and disable emit on fail.

For some reason I put it in goroutine (which makes no sense). I guess it
was time pressure and I actually meant to init variable and truth value
functionally. But ended up introducing a race condition (which actually
could not result in runtime error in practice).

Still a hack. But I don't have time or motivation to rewrite the shipper
code into a proper handler for now.